+++ workshop/Contents/mods/InventoryTags/common/media/lua/shared/InventoryTags/InventoryTags_Core.lua
@@ -1,6 +1,6 @@
 InventoryTags = InventoryTags or {}
 local IT = InventoryTags
-IT.VERSION = "0.1.2-SELECT3"
+IT.VERSION = "0.1.3"
 IT.PROTOCOL = 6
 IT.ACCEPT_FUNCTION = "InventoryTags.Filter.acceptItem"
 IT.MAX_AUTO_ORGANIZE_STEPS = 500